+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/DefaultLookAndFeelSelectionDialogBuilder.java
@@ -86,7 +86,8 @@ final class DefaultLookAndFeelSelectionDialogBuilder implements LookAndFeelSelec
 		basePanel.setBorder(BorderFactory.createEmptyBorder(PADDING, PADDING, 0, PADDING));
 		basePanel.add(lookAndFeelComboBox, BorderLayout.CENTER);
 		if (auxiliaryLookAndFeelsAvailable()) {
-			basePanel.add(PanelBuilder.builder(new FlowLayout(FlowLayout.TRAILING))
+			basePanel.add(PanelBuilder.builder()
+							.layout(new FlowLayout(FlowLayout.TRAILING))
 							.add(CheckBoxBuilder.builder()
 											.link(lookAndFeelComboBox.includeInstalled())
 											.text(MESSAGES.getString("include_installed"))